♻️ refactor: refactor the global Store
