+++ src/app/chat/features/SettingButton.tsx
@@ -5,10 +5,10 @@ import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { DESKTOP_HEADER_ICON_SIZE, MOBILE_HEADER_ICON_SIZE } from '@/const/layoutTokens';
+import { useGlobalStore } from '@/store/global';
+import { SidebarTabKey } from '@/store/global/initialState';
 import { useSessionStore } from '@/store/session';
 import { sessionSelectors } from '@/store/session/selectors';
-import { useUserStore } from '@/store/user';
-import { SidebarTabKey } from '@/store/user/initialState';
 import { pathString } from '@/utils/url';
 
 const SettingButton = memo<{ mobile?: boolean }>(({ mobile }) => {
@@ -21,7 +21,7 @@ const SettingButton = memo<{ mobile?: boolean }>(({ mobile }) => {
       icon={AlignJustify}
       onClick={() => {
         if (isInbox) {
-          useUserStore.setState({
+          useGlobalStore.setState({
             sidebarKey: SidebarTabKey.Setting,
           });
           router.push('/settings/agent');
