+++ src/app/settings/_layout/Desktop/Header.tsx
@@ -7,7 +7,7 @@ import { useTranslation } from 'react-i18next';
 import { Flexbox } from 'react-layout-kit';
 
 import { useActiveSettingsKey } from '@/hooks/useActiveSettingsKey';
-import { SettingsTabs } from '@/store/user/initialState';
+import { SettingsTabs } from '@/store/global/initialState';
 
 const Header = memo(() => {
   const { t } = useTranslation('setting');
