+++ src/app/settings/features/UpgradeAlert.tsx
@@ -5,10 +5,10 @@ import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { MANUAL_UPGRADE_URL } from '@/const/url';
-import { useUserStore } from '@/store/user';
+import { useGlobalStore } from '@/store/global';
 
 const UpgradeAlert = memo(() => {
-  const [hasNewVersion, latestVersion] = useUserStore((s) => [s.hasNewVersion, s.latestVersion]);
+  const [hasNewVersion, latestVersion] = useGlobalStore((s) => [s.hasNewVersion, s.latestVersion]);
   const { t } = useTranslation('common');
 
   return (
