+++ src/layout/DefaultLayout/Desktop/SideBar/index.tsx
@@ -1,7 +1,7 @@
 import { SideNav } from '@lobehub/ui';
 import { memo } from 'react';
 
-import { SidebarTabKey } from '@/store/user/initialState';
+import { SidebarTabKey } from '@/store/global/initialState';
 
 import Avatar from './Avatar';
 import BottomActions from './BottomActions';
