+++ src/layout/GlobalProvider/StoreInitialization.tsx
@@ -7,6 +7,7 @@ import { createStoreUpdater } from 'zustand-utils';
 import { useIsMobile } from '@/hooks/useIsMobile';
 import { useEnabledDataSync } from '@/hooks/useSyncData';
 import { useAgentStore } from '@/store/agent';
+import { useGlobalStore } from '@/store/global';
 import { useUserStore } from '@/store/user';
 
 const StoreInitialization = memo(() => {
@@ -15,9 +16,13 @@ const StoreInitialization = memo(() => {
     s.useFetchUserConfig,
     s.useInitPreference,
   ]);
+  const useInitGlobalPreference = useGlobalStore((s) => s.useInitGlobalPreference);
+
   const useFetchDefaultAgentConfig = useAgentStore((s) => s.useFetchDefaultAgentConfig);
   // init the system preference
   useInitPreference();
+  useInitGlobalPreference();
+
   useFetchDefaultAgentConfig();
 
   const { isLoading } = useFetchServerConfig();
@@ -25,7 +30,7 @@ const StoreInitialization = memo(() => {
 
   useEnabledDataSync();
 
-  const useStoreUpdater = createStoreUpdater(useUserStore);
+  const useStoreUpdater = createStoreUpdater(useGlobalStore);
 
   const mobile = useIsMobile();
   const router = useRouter();
